# lab4/app/Models/ProductCategory.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
/**
 * @OA\Schema (
 *  title="ProductCategory",
 *  description="ProductCategory model",
 *  @OA\Property(
 *      property="id",
 *      type="integer"
 *  ),
 *  @OA\Property(
 *      property="name",
 *      type="string"
 *  ),
 *  @OA\Property(
 *      property="created_at",
 *      type="date"
 *  ),
 *  @OA\Property(
 *      property="updated_at",
 *      type="date"
 *  ),
 * )
 */
class ProductCategory extends Model
{
    use HasFactory;

    protected $fillable = [
      'name'
    ];

    public function products()
    {
        return $this->hasMany(Product::class, 'category_id');
    }
}