# lab4/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PizzGram Api Documentation",
        "description": "L5 Swagger PizzGram Api description",
        "contact": {
            "email": "eliseevv02@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "PizzGram API Server"
        }
    ],
    "paths": {
...
        "/product_categories": {
            "get": {
                "tags": [
                    "Product categories"
                ],
                "summary": "Get list of all product categories",
                "description": "Returns list of product categories",
                "operationId": "getAllProductCategories",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product categories"
                ],
                "summary": "Add new product category",
                "description": "Adds new product category",
                "operationId": "addProductCategory",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Product category name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The adding request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Some Product Categoty"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The product category was created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product_category": {
                                            "$ref": "#/components/schemas/ProductCategory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/product_categories/{id}": {
            "get": {
                "tags": [
                    "Product categories"
                ],
                "summary": "Get one product category by id",
                "description": "Returns object of product category",
                "operationId": "getProductCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product_category": {
                                            "$ref": "#/components/schemas/ProductCategory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product categories"
                ],
                "summary": "Update existing prodcut category",
                "description": "Returns updated prodcut category data",
                "operationId": "updateProductCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Product category name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The updating request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Some Product Categoty"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The product category was updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductCategory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product categories"
                ],
                "summary": "Delete product category by id",
                "description": "Deletes product category by id",
                "operationId": "deleteProductCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product category was deleted."
                    },
                    "401": {
                        "description": "Unauthorised"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductCategory": {
                "title": "ProductCategory",
                "description": "ProductCategory model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}